requirements.txt
fastapi
uvicorn[standard]
openai-whisper
ffmpeg-python
langchain
langchain-core
langchain-openai
pydantic
pydantic-settings
python-multipart
loguru
fastapi[all]
